fix bugs

+++ utils/replay_buffer.py
@@ -73,7 +73,7 @@ class ReplayBufferDiscreteAction(ReplayBuffer):
             s, a, s_, r, dw, done = zip(*batch)
 
         s = torch.tensor(np.asarray(s), dtype=torch.float)
-        a = torch.tensor(np.asarray(a))
+        a = torch.tensor(np.asarray(a), dtype=torch.int64)
         if with_log:
             a_logprob = torch.tensor(np.asarray(a_logprob), dtype=torch.float)
         s_ = torch.tensor(np.asarray(s_), dtype=torch.float)
